+++ i7dw/io.py
@@ -1,14 +1,14 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
 import bisect
 import os
+import json
 import pickle
 import shutil
 import sqlite3
 import struct
 import zlib
-from multiprocessing import Pool, Process, Queue
+from multiprocessing import Pool, Queue
 from tempfile import mkdtemp, mkstemp
 from typing import Any, Callable, Generator, Iterable, Tuple, Union
 
@@ -591,3 +591,38 @@ class TempFile(object):
             os.remove(self.path)
         except FileNotFoundError:
             pass
+
+
+class JsonFileOrganizer(object):
+    def __init__(self, root: str, items_per_file: int=10000,
+                 files_per_dir: int=1000):
+        self.root = root  # must already exist
+        self.dir = root
+        self.count = 0
+        self.items_per_file = items_per_file
+        self.files_per_dir = files_per_dir
+        self.items = []
+
+    def add(self, item):
+        self.items.append(item)
+
+        if len(self.items) == self.items_per_file:
+            self.flush()
+
+    def flush(self):
+        if not self.items:
+            return
+        elif self.count + 1 == self.files_per_dir:
+            # Too many files in directory: create a subdirectory
+            self.dir = mkdtemp(dir=self.dir)
+            self.count = 0
+
+        fd, path = mkstemp(dir=self.dir)
+        os.close(fd)
+
+        with open(path, "wt") as fh:
+            json.dump(self.items, fh)
+
+        self.count += 1
+        self.items = []
+        os.rename(path, path + ".json")
